Fix for room setup for MixedRealitySample SpatialSDK
Summary:
This diff is for the github change as the issue exists in 0.8.0

When the user selects to set up room in MixedRealitySample, the system saves the room data to device storage but room data never loads. The condition would resolve itself on app restart as onCreate() -> loadSceneFromDevice() runs

Now when the user selects to set up room, the loadSceneFromDevice() added immediately loads the new room data so the user immediately sees the room setup with the colliders and expected boundaries

the requestSceneCapture() was saving the room data to the system but there was no 'load' when in app, adding the loadSceneFromDevice() bridges the gap so the saved data automatically loads after scene capture completes.

Reviewed By: bleavitt24

Differential Revision: D85466038

fbshipit-source-id: 33e06630857a3b715decb861a64ec91978b8b532