Fix youtube embeds
Summary:
Found this discussion on reddit:
https://www.reddit.com/r/ObsidianMD/comments/1ogzv1s/error_153_on_all_my_embedded_youtube_videos_all/

Google's Documentation says we need to attach a referrer to the request:
https://developers.google.com/youtube/terms/required-minimum-functionality#embedded-player-api-client-identity

"""
Within the URL, the domain name must be your application identifier ("app ID") that is registered with the store that distributed your app to end users. If your app is provided to users through an alternative distribution channel, use the app ID that is registered with the operating system during app install. In most cases, your app ID will be a reversed domain name (also known as "reverse-DNS format"), like com.google.android.youtube.
"""

This diff

Reviewed By: bleavitt24

Differential Revision: D86109439

fbshipit-source-id: 72d8372cf876ae6b85b033abc7d9d2e3f140cfbf